pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "musiol-rag"
version = "0.1.0"
description = "A modular Retrieval-Augmented Generation system"
readme = "README.md"
authors = [{ name = "Martin Musiol" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "sentence-transformers>=2.2.2",
    "faiss-cpu>=1.7.4",
    "pydantic>=1.9.0",
    "sqlalchemy>=1.4.23",
    "asyncpg>=0.25.0",
    "uvicorn>=0.15.0",
    "python-dotenv>=1.0.0",
    "openai>=1.3.0",
    "numpy>=1.24.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-asyncio>=0.14.0",
    "black>=21.0",
    "isort>=5.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["musiol_rag*"]

[tool.black]
line-length = 88
target-version = ["py38"] 